requirements.txt
apache-airflow-providers-snowflake
snowflake-connector-python
pandas
yfinance
statsmodels
